refactor reference_counted: add weak
