+++ src/containers/reference_counted/CMakeLists.txt
@@ -1,6 +1,6 @@
 target_sources(${PROJECT_NAME}
     PRIVATE
-        ictrl_block.c
-
         reference_counted.h
 )
+
+add_subdirectory(ctrl_block)
