+++ src/containers/reference_counted/ctrl_block/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(${PROJECT_NAME}
+    PRIVATE
+        ictrl_block.c
+        ictrl_block.h
+)
