+++ src/containers/reference_counted/ctrl_block/combined_ctrl_block/combined_ctrl_block.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "combined_ctrl_block_allocator.h"
-#include "ictrl_block.h"
+#include <containers/reference_counted/ctrl_block/ictrl_block.h>
 #include "combined_ctrl_block_common.h"
 #include <memory/common.h>
