+++ src/containers/reference_counted/ctrl_block/separated_ctrl_block/separated_ctrl_block_common.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "common.h"
+#include <containers/reference_counted/common.h>
 #include <memory/common.h>
 #include <memory/allocator/allocator.h>
